Use hand-coded lexer requiring SPDX list packages
Prior to this, `npm prepublish` generated both lexer and parser
using Jison, hard-coding the versions of `spdx-license-ids` and
`spdx-exceptions` installed as `devDependencies` on the publisher's
machine into the package.  There were no `dependencies`.

This commit adds a hand-coded lexer, and uses Jison to generate only
the parser.  As a result, the SPDX list-data packages become regular
`dependencies`, `npm install` can install the latest versions,
and the lexer can `require()` them.  This will help ensure that
those relying on `spdx-expression-parse` get the latest SPDX lists,
without waiting for `spdx-expression-parse` to republish.