Drop CC-BY-3.0 from npm license expression
Before version 2.0.0, releases of this package required the
`spdx-license-ids` and `spdx-exceptions` packages as development
dependencies only, and used them to generate a lexer.  The generated
lexer contained transformed copies of the lists.

It wasn't legally clear whether lists---arrays of string
identifiers---embodied sufficient creativity to trigger copyright
protection under all possible copyright laws of interest to users.
That being the case, and out of an abundance of caution, the `license`
property in this package's `package.json` file mentioned the license
from The Linux Foundation, `CC-BY-3.0`, as well as the license for
the code implementing the parser, `MIT`.

Beginning in the 2.0.0 release, this package depends on
`spdx-license-ids` and `spdx-exceptions` as regular npm
dependencies. Downstream users installing this package also download
the list packages.  This package `require()`s them, rather than
bundling their content in generated code.

Accordingly, users reviewing license data should now look to the
`package.json` files for the individual list packages for information
on their licensing terms.  This package's `license` expression can
express only the terms that apply to the code it contains.