Handwritten parser
This subjective and opinionated patch fixes many issues but has
several drawbacks.

The parser should be usable in strict mode now (see #9). It is not
longer generated by `jison`. `parse.js` contains a simple handwritten
recursive decent parser. It should be easier to debug and to hack, but
there is definitely more code and one can prefer parser generators.
Note that `jison` seems to have big maintainance issues.

Note that [Yargs](https://github.com/yargs/yargs) is not usable in strict mode because it depends on
this package.

I also simplified the lexer (IMHO).

I wrote more Mocha-based tests in `test/index.js`. I really like the
embedded tests in `README.md`, but I was afraid to bloat it with
boring spec-compliance tests.

This patch does not introduce a breaking change and does not change
dependencies (however, it changes devDependencies).